FireBall change to SandBall

+++ Assets/Scripts/Missile/FireballMissile.cs
@@ -50,6 +50,6 @@ public class fireballMissile : MonoBehaviour
     private void setAnimatorDestroy(int animatorDestroyIndex)
     {
         animator.SetTrigger("Destroy");
-        animator.SetInteger("DestroyAnimIndex", animatorDestroyIndex);
+        animator.SetFloat("DestroyAnimIndex", animatorDestroyIndex);
     }
 }
